models complete

+++ db/migrate/20230216023159_add_rating_to_movies.rb
@@ -1,5 +1,5 @@
 class AddRatingToMovies < ActiveRecord::Migration[7.0]
   def change
-    add_column :movies, :rating, :integer
+    add_column :movies, :rating, :float
   end
 end
